+++ test/game-ko-test.js
@@ -102,6 +102,7 @@ describe("ko restriction", function() {
       game.playAt(3, 8); // b tenuki
       game.playAt(2, 0); // w
       game.playAt(4, 8); // b tenuki -- (*)
+
       game.playAt(0, 2); // w
 
       expect(game.playAt(0, 0)).to.be.true; // b
@@ -161,6 +162,108 @@ describe("ko restriction", function() {
       expect(game.currentPlayer()).to.equal("white");
       expect(game.currentState().playedPoint.y).to.equal(4);
       expect(game.currentState().playedPoint.x).to.equal(4);
+    });
+  });
+
+  describe("with situational superko rules", function() {
+    it("allows repeating a previous position if a different player is to play", function() {
+      var game = new Game({ koRule: "situational-superko" });
+
+      // cycle for ● repeatedly losing 2 stones
+      // ┌─●─┬─○─●
+      // ●─○─○─○─●
+      // ●─┼─○─●─●
+      // ○─○─○─●─┼
+      // ●─●─●─●─┼
+      game.playAt(0, 3); // b
+      game.playAt(0, 4); // w
+      game.playAt(1, 3); // b
+      game.playAt(1, 4); // w
+      game.playAt(1, 2); // b
+      game.playAt(2, 4); // w
+      game.playAt(1, 1); // b
+      game.playAt(2, 3); // w
+      game.playAt(2, 2); // b
+      game.playAt(3, 3); // w
+      game.playAt(3, 2); // b
+      game.playAt(4, 3); // w
+      game.playAt(3, 1); // b
+      game.playAt(4, 2); // w
+      game.playAt(3, 0); // b
+      game.playAt(4, 1); // w
+      game.playAt(0, 8); // b tenuki
+      game.playAt(4, 0); // w
+      game.playAt(1, 8); // b tenuki
+      game.playAt(0, 1); // w
+      game.playAt(2, 8); // b tenuki
+      game.playAt(1, 0); // w
+      game.playAt(3, 8); // b tenuki
+      game.playAt(2, 0); // w
+      game.playAt(4, 8); // b tenuki -- (*)
+
+      game.playAt(0, 2); // w
+
+      expect(game.playAt(0, 0)).to.be.true; // b
+
+      // this move is allowed with situational superko since it repeats (*),
+      // but (*) was created by black's tenuki, and it's white recreating (*)
+      // with this move.
+      expect(game.playAt(0, 1)).to.be.true; // w
+      expect(game.intersectionAt(0, 1).value).to.equal("white");
+      expect(game.currentPlayer()).to.equal("black");
+      expect(game.currentState().playedPoint.y).to.equal(0);
+      expect(game.currentState().playedPoint.x).to.equal(1);
+    });
+  });
+
+  describe("with natural situational superko rules", function() {
+    it("allows repeating a previous situation if that situation was created by a pass", function() {
+      [
+        ["positional-superko", false],
+        ["situational-superko", false],
+        ["natural-situational-superko", true]
+      ].forEach(([rule, expectedResult]) => {
+        var game = new Game({ koRule: rule, boardSize: 5 });
+
+        // ┌─●─○─┬─┐
+        // ●─○─○─┼─○
+        // ●─●─○─○─┤
+        // ├─●─●─○─○
+        // ●─┴─●─○─┘
+
+        game.playAt(0, 2); // b
+        game.playAt(1, 0); // w
+        game.playAt(1, 2); // b
+        game.playAt(2, 0); // w
+        game.playAt(1, 1); // b
+        game.playAt(2, 1); // w
+        game.playAt(2, 2); // b
+        game.playAt(3, 1); // w
+        game.playAt(2, 3); // b
+        game.playAt(3, 2); // w
+        game.playAt(3, 3); // b
+        game.playAt(4, 2); // w
+        game.playAt(1, 4); // b
+        game.playAt(4, 0); // w
+        game.playAt(3, 4); // b
+        game.playAt(0, 1); // w
+        game.playAt(4, 3); // b
+
+        game.pass(); // w (*)
+        game.playAt(0, 0); // b -- take ko
+
+        // this white move recreates the position at (*).
+        // that position was created by pass by white.
+        //
+        //   * under situational superko rules, this white move is
+        //     illegal, becuase white is recreating a position (*)
+        //     that white created.
+        //   * under natural situational superko rules, this white move
+        //     is legal, because despite recreating (*), (*) was created
+        //     with a pass, and not a board play.
+        //
+        expect(game.playAt(0, 1), `under ${rule}, expected playing at 0,1 to be ${expectedResult ? "valid" : "illegal"}`).to.equal(expectedResult);
+      })
     });
   });
 });